+++ sphinx/source/changelog.rst
@@ -23,6 +23,8 @@ Changelog
   of the tape's tones, pulse sequences and data blocks)
 * Added the ``--tape-skip`` option to :ref:`tapinfo.py` (for skipping one or
   more blocks on a tape)
+* Fixed how the ``--basic`` option of :ref:`tapinfo.py` handles a block that
+  has no data
 * Fixed how two base prefixes are handled when applied to an instruction with
   one operand
 * Fixed the parsing of a 48K SNA snapshot that has SP set to 0xFFFF